More work on Palette16ChunkFormat

+++ terra-core/src/main/java/com/ritualsoftheold/terra/material/TerraMaterial.java
@@ -37,7 +37,7 @@ public class TerraMaterial {
         return worldId;
     }
     
-    void setWorldId(short id) {
+    void setWorldId(int id) {
         worldId = id;
     }
